refactor: consolidate all RestartGame tests to ConstructorsAndRestartGame test class, rename RestartRehide test class to Rehide
